Remove legacy stuff from toast interface
